Move action persistence models to individual files
