+++ st2actions/tests/integration/test_resultstracker.py
@@ -1,7 +1,8 @@
 import eventlet
 
 import st2actions.resultstracker as results_tracker
-from st2common.persistence.action import LiveAction, ActionExecutionState
+from st2common.persistence.executionstate import ActionExecutionState
+from st2common.persistence.liveaction import LiveAction
 from st2tests.base import (DbTestCase, EventletTestCase)
 from st2tests.fixturesloader import FixturesLoader
