+++ st2common/st2common/persistence/actionalias.py
@@ -13,26 +13,12 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from st2common.models.db.action import action_access
+from st2common.models.db.action import actionalias_access
 from st2common.persistence import base as persistence
-from st2common.persistence.actionalias import ActionAlias
-from st2common.persistence.execution import ActionExecution
-from st2common.persistence.executionstate import ActionExecutionState
-from st2common.persistence.liveaction import LiveAction
-from st2common.persistence.runner import RunnerType
-
-__all__ = [
-    'Action',
-    'ActionAlias',
-    'ActionExecution',
-    'ActionExecutionState',
-    'LiveAction',
-    'RunnerType'
-]
 
 
-class Action(persistence.ContentPackResource):
-    impl = action_access
+class ActionAlias(persistence.Access):
+    impl = actionalias_access
 
     @classmethod
     def _get_impl(cls):
