+++ st2common/st2common/persistence/executionstate.py
@@ -13,27 +13,24 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from st2common.models.db.action import action_access
+from oslo.config import cfg
+
+from st2common import transport
+from st2common.models.db.action import actionexecstate_access
 from st2common.persistence import base as persistence
-from st2common.persistence.actionalias import ActionAlias
-from st2common.persistence.execution import ActionExecution
-from st2common.persistence.executionstate import ActionExecutionState
-from st2common.persistence.liveaction import LiveAction
-from st2common.persistence.runner import RunnerType
-
-__all__ = [
-    'Action',
-    'ActionAlias',
-    'ActionExecution',
-    'ActionExecutionState',
-    'LiveAction',
-    'RunnerType'
-]
 
 
-class Action(persistence.ContentPackResource):
-    impl = action_access
+class ActionExecutionState(persistence.Access):
+    impl = actionexecstate_access
+    publisher = None
 
     @classmethod
     def _get_impl(cls):
         return cls.impl
+
+    @classmethod
+    def _get_publisher(cls):
+        if not cls.publisher:
+            cls.publisher = transport.actionexecutionstate.ActionExecutionStatePublisher(
+                cfg.CONF.messaging.url)
+        return cls.publisher
