+++ st2common/tests/unit/test_action_api_validator.py
@@ -22,7 +22,7 @@ import mock
 
 from st2common.exceptions.apivalidation import ValueValidationException
 from st2common.models.api.action import (ActionAPI, RunnerTypeAPI)
-from st2common.persistence.action import RunnerType
+from st2common.persistence.runner import RunnerType
 import st2common.validators.api.action as action_validator
 from st2tests import DbTestCase
 from st2tests.fixtures import executions as fixture
